add,update,deleteのバグを改善させた

+++ KUMA/app/src/main/java/com/seven_dream/kuma/ScheduleOneday.kt
@@ -4,6 +4,7 @@ import android.content.Intent
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.TextView
+import android.widget.Toast
 import kotlinx.android.synthetic.main.activity_schedule_oneday.*
 import com.seven_dream.kuma.R.layout.activity_schedule_oneday
 
@@ -30,7 +31,23 @@ class ScheduleOneday : AppCompatActivity() {
         /* addボタンをクリックすると構造体NewPlanにEditTextで入力した値を格納し、
          データベース内のテーブルに構造体を格納する*/
         btn_add.setOnClickListener {
-            val plans = NewPlan (
+            if (db.isRecord(Integer.parseInt(edt_id.text.toString())) == false) {
+                val plans = NewPlan (
+                    Integer.parseInt(edt_id.text.toString()),
+                    edt_date.text.toString(),
+                    edt_title.text.toString(),
+                    edt_timeBegin.text.toString(),
+                    edt_timeEnd.text.toString(),
+                    edt_place.text.toString(),
+                    edt_memo.text.toString()
+                )
+                db.addPlans(plans)
+                Toast.makeText(this, "登録されました!", Toast.LENGTH_SHORT).show()
+                finish()
+            } else {
+                Toast.makeText(this, "すでに登録されているIDです!", Toast.LENGTH_SHORT).show()
+            }
+            /*val plans = NewPlan (
                 Integer.parseInt(edt_id.text.toString()),
                 edt_date.text.toString(),
                 edt_title.text.toString(),
@@ -40,12 +57,13 @@ class ScheduleOneday : AppCompatActivity() {
                 edt_memo.text.toString()
             )
             db.addPlans(plans)
-            finish()
+            finish()*/
             //refreshData()
         }
 
         //指定したIDを元にデータベース内のテーブルの内容を変更する
         btn_update.setOnClickListener {
+            if (db.isRecord(Integer.parseInt(edt_id.text.toString())) == true) {
             val plans = NewPlan (
                 Integer.parseInt(edt_id.text.toString()),
                 edt_date.text.toString(),
@@ -55,24 +73,33 @@ class ScheduleOneday : AppCompatActivity() {
                 edt_place.text.toString(),
                 edt_memo.text.toString()
             )
-            db.updatePlans(plans)
-            finish()
+                db.updatePlans(plans)
+                Toast.makeText(this, "変更されました", Toast.LENGTH_SHORT).show()
+                finish()
+            } else {
+                Toast.makeText(this, "まだ登録されていないIDです!", Toast.LENGTH_SHORT).show()
+            }
             //refreshData()
         }
 
         //指定したIDを元にデータベース内のテーブルの内容を削除する
         btn_delete.setOnClickListener {
-            val plans = NewPlan (
-                Integer.parseInt(edt_id.text.toString()),
-                edt_date.text.toString(),
-                edt_title.text.toString(),
-                edt_timeBegin.text.toString(),
-                edt_timeEnd.text.toString(),
-                edt_place.text.toString(),
-                edt_memo.text.toString()
-            )
-            db.deletePlans(plans)
-            finish()
+            if (db.isRecord(Integer.parseInt(edt_id.text.toString())) == true) {
+                val plans = NewPlan(
+                    Integer.parseInt(edt_id.text.toString()),
+                    edt_date.text.toString(),
+                    edt_title.text.toString(),
+                    edt_timeBegin.text.toString(),
+                    edt_timeEnd.text.toString(),
+                    edt_place.text.toString(),
+                    edt_memo.text.toString()
+                )
+                db.deletePlans(plans)
+                Toast.makeText(this, "削除されました", Toast.LENGTH_SHORT).show()
+                finish()
+            } else {
+                Toast.makeText(this, "まだ登録されていないIDです!", Toast.LENGTH_SHORT).show()
+            }
             //refreshData()
         }
